fix: fix issues introduced by rebase
