+++ packages/dendron-viz/src/processDendronNotes.ts
@@ -1,8 +1,10 @@
 import {
-  DEngineClient,
   DNodeUtils,
+  NoteDictsUtils,
+  NoteFnameDictUtils,
   DVault,
   NoteProps,
+  NotePropsByIdDict,
 } from "@dendronhq/common-all";
 import * as nodePath from "path";
 import { shouldExcludePath } from "./shouldExcludePath";
@@ -14,14 +16,14 @@ function isDir(note: NoteProps) {
 
 export const processDir = async ({
   rootPath = "",
-  engine,
+  notes: notesById,
   vault,
   excludedPaths = [],
   excludedGlobs = [],
 }: {
   rootPath: string;
   vault: DVault;
-  engine: DEngineClient;
+  notes: NotePropsByIdDict;
   excludedPaths?: string[];
   excludedGlobs?: string[];
 }): Promise<FileType> => {
@@ -33,18 +35,26 @@ export const processDir = async ({
 
   /* Given a file name, get corresponding Dendron note */
   async function getNote(fname: string): Promise<NoteProps> {
-    const note = await engine.findNotes({ fname, vault });
+    const notesByFname =
+      NoteFnameDictUtils.createNotePropsByFnameDict(notesById);
+    const notes = NoteDictsUtils.findByFname(
+      fname,
+      { notesById, notesByFname },
+      vault
+    );
+
+    const note = notes[0];
 
     if (note === undefined) {
       throw new Error(`Issue trying to find the note ${fname}`);
     }
 
-    return note[0];
+    return note;
   }
 
   /* Given a note, get its child notes */
   function getChildren(note: NoteProps): NoteProps[] {
-    return note.children.map((id) => engine.notes[id]);
+    return note.children.map((id) => notesById[id]);
   }
 
   /* Given a note, get file stats needed for Tree React component */
